Update: Added Party paramter for debugging.
The party paramter is extra data for reserved positions that helps identitfy the who of a UI reservation.

Update:
Search for proccessing time is also more thorough.

Fix:
getNewPos no longer reserves the inital position. This fixes the UI issue where it seemed like the recipe was duplicating reservations. It would cause the placement of the first position to be shifted to avoid the unintented reservation.

Tweak:
Clarifyied jsonToItem logging

Update:
Recipe Nuggets are now correctly added. Chances and amounts are also respected. Cleaned up the extra math to determine the experience.

Added logging for when recipe nuggets or when experience and helpful comment.

Update:
Proccesing time is now respected.

Tweak: Log on the for each recipe has been sperated and simplified.

TODO: Handling recipes that have the same input

Tweak:
Lowered amount of recalulations by a staggring amount inside of the core recipe function

Notes: Debug recipes are still active.