Lowered Priority to catch KubeJS recipes (untested)
Block, fluid, or Item tags will now all correctly parsed to items to be used in the recipe for cores.

Core recipes are auto generated based on the core blocks that will be used inside the actual block. Ensuring a player has to have more of a reliable source of the core by requiring two uses of the core item. One, in the recipe, and one in the machine.