A recipe for the nether star, the various shadow and refined radiance casings have been added using the createcompounds mod.
Uses for the shadow steel, the refined radiance, and the chromatic compound have been added as well.
Recipe difficulty is high.

Random chances for the duplication ensure automation is mandatory for the true benefit. Thus, making the recipes more resource intensive in what isn't given back. A feedback system is also required. All this to hopefully penalize players that haven't respected the other aspects of automation to get a recipe for the netherstar.